% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{atomicSim}
\alias{atomicSim}
\title{atomic realization for parallel simulation study}
\usage{
atomicSim(n, lambda_0, gamma, theta, eta = 1, mu = 1, s)
}
\arguments{
\item{n}{Number of samples (effective arrivals) to generate.}

\item{lambda_0}{The constant arrival rate}

\item{gamma}{The periodic arrival rate maximum amplitude}

\item{theta}{Parameter of the exponential patience}

\item{eta}{Shape parameter of job size.}

\item{mu}{Rate parameter of job size.}

\item{s}{Number of servers.}
}
\value{
A vector with Boris's etimates and Liron's etsimates
}
\description{
Atomic Simulation for periodic arrivals (cosine model)
}
