% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{customerExp}
\alias{customerExp}
\title{Customer's job size and patience
Provides with a realization of Y~Exp(theta) and B~Gamma(eta,mu)}
\usage{
customerExp(theta, eta, mu)
}
\arguments{
\item{theta}{The parameter of the exponential patience.}

\item{eta}{Shapae parameter of the job size.}

\item{mu}{rate parameter of the job size.}
}
\value{
A list with elements 'Patience' and 'Jobsize'
}
\description{
Customer's job size and patience
Provides with a realization of Y~Exp(theta) and B~Gamma(eta,mu)
}
