% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{evaluateGridFromRealization}
\alias{evaluateGridFromRealization}
\title{Evaluate the log likelihood function over a grid of parameter values}
\usage{
evaluateGridFromRealization(dati, grid)
}
\arguments{
\item{dati}{a dataset A,W,X}

\item{grid}{a grid of parameters}
}
\value{

}
\description{
Evaluate the log likelihood function over a grid of parameter values
}
