% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gridFromFilePath}
\alias{gridFromFilePath}
\title{Evaluate a parameter grid's likelihood from AWX file}
\usage{
gridFromFilePath(path, grid)
}
\arguments{
\item{path}{a AWX.csv file path}

\item{grid}{output of makeParGrid()}
}
\value{

}
\description{
Evaluate a parameter grid's likelihood from AWX file
}
