% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ithLL}
\alias{ithLL}
\title{Compute log-likelihood from a dataset A,W,X}
\usage{
ithLL(AWX, gamma, lambda_0, theta)
}
\arguments{
\item{AWX}{dataframe with columns A,W,X}
}
\value{

}
\description{
Compute log-likelihood from a dataset A,W,X
}
\details{
this auxiliary function is used within `meanLogLikelihoodFromList`.
}
