% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeParGrid}
\alias{makeParGrid}
\title{Make parameter grid for the average likelihood computation}
\usage{
makeParGrid(params, spans, grid.sizes)
}
\arguments{
\item{params}{named list of the parameters (gamma,lambda_0,theta)}

\item{spans}{vector of length 3 which determines the span of each axis in the
grid. Values must be in the interval [0,1). See details.}

\item{grid.sizes}{an integer vector of length 1 or 3. In the former case, all
axes have the same number of points, and in the latter each parameter is
matched to the corresponding element of `grid.sizes`.}
}
\value{

}
\description{
Make parameter grid for the average likelihood computation
}
