% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeSimFilesAWX}
\alias{makeSimFilesAWX}
\title{Function to make many files with simulation results}
\usage{
makeSimFilesAWX(
  path,
  n.cores = parallel::detectCores() - 2,
  N_files,
  n_obs,
  gamma,
  lambda_0,
  theta,
  s,
  eta,
  mu
)
}
\arguments{
\item{path}{Absolute path of the folder to create the files in (will be set to wd)}

\item{N_files}{Desired number of files to be created in the working directory}

\item{n_obs}{Number of effective arrivals per file. Defaults to 10,000. Do not use less than 5000.}

\item{gamma}{The periodic arrival rate maximum amplitude}

\item{lambda_0}{The constant arrival rate}

\item{theta}{Parameter of the exponential patience}

\item{s}{Number of servers.}

\item{eta}{Shape parameter of job size.}

\item{mu}{Rate parameter of job size.}
}
\value{

}
\description{
Function to make many files with simulation results
}
\details{
The filenames have the parameter values encoded alongside a timestamp,
which is meant for protection against overwriting in case of repeated calls to this function.
Also note that if files take less than a second to generate, they will not be written.
}
