% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meanLogLikelihoodFromList}
\alias{meanLogLikelihoodFromList}
\title{Average log likelihood from many files}
\usage{
meanLogLikelihoodFromList(gamma, lambda_0, theta, res_list)
}
\arguments{
\item{res_list}{a list where each element is a dataframe with columns A,W,X}
}
\value{
the average likelihood for the parameter values provided
}
\description{
Average log likelihood from many files
}
\examples{
# use with vectorize:
VmLLVec <- Vectorize(meanLogLikelihoodFromList,vectorize.args = c("gamma","lambda_0","theta"))

}
