% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mleBoris}
\alias{mleBoris}
\title{MLE for the cosine arrival + exponential patience model}
\usage{
mleBoris(dati, PARAMS)
}
\arguments{
\item{PARAMS}{(temporary) True values of parameters to use a starting points}

\item{type}{type of log-likelihood to be optimized - sum ("s") or mean ("m")}
}
\value{
gradient vector of the negative log-likelihood
}
\description{
MLE for the cosine arrival + exponential patience model
}
