% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mleFromFilePath}
\alias{mleFromFilePath}
\title{Evaluate MLE's (Boris and Liron) from AWX file}
\usage{
mleFromFilePath(path)
}
\arguments{
\item{path}{a AWX.csv file path}

\item{grid}{output of makeParGrid()}
}
\value{
Length 5 vector - 3 Boris + 2 Liron
}
\description{
Evaluate MLE's (Boris and Liron) from AWX file
}
