% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mleKnownArrival}
\alias{mleKnownArrival}
\title{MLE for theta provided gamma and lambda_0}
\usage{
mleKnownArrival(AWX, params)
}
\arguments{
\item{AWX}{data}

\item{params}{c(gamma,lambda_0)}
}
\value{
MLE
}
\description{
MLE for theta provided gamma and lambda_0
}
