% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mleLironThetaLambda}
\alias{mleLironThetaLambda}
\title{Liron MLE}
\usage{
mleLironThetaLambda(AWX, acc = 1e-04)
}
\arguments{
\item{acc}{}
}
\value{
A vector of theta and lambda estimates
}
\description{
Liron MLE
}
