% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{negLogLikelihoodMean}
\alias{negLogLikelihoodMean}
\title{(negative) mean log-likelihood for the sinusoidal arrivals model
The mean is returned instead of the sum - should help gradient based optimizers}
\usage{
negLogLikelihoodMean(params, AWX)
}
\arguments{
\item{params}{A vector with values (gamma,lambda_0, theta).}

\item{AWX}{compact simulation results for memory economy (A,W,X).}
}
\value{
The negative log-likelihood at the point provided.
}
\description{
(negative) mean log-likelihood for the sinusoidal arrivals model
The mean is returned instead of the sum - should help gradient based optimizers
}
