% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nextArrivalCosine}
\alias{nextArrivalCosine}
\title{Next Arrival of inhomogeneous Poisson process}
\usage{
nextArrivalCosine(current_time, gamma, lambda_0)
}
\arguments{
\item{current_time}{The current time (a non-negative real number)}

\item{gamma}{Parameter of the cosine arrival}

\item{lambda_0}{Parameter of the cosine arrival}
}
\value{
The time of the next arrival.
}
\description{
Generates the next arrival time (not inter-arrival!) of a nonhomogeneous Poisson process.
}
