% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pltQueueByHour}
\alias{pltQueueByHour}
\title{Plot the patience distribution}
\usage{
pltQueueByHour(RES, n_quantiles = 4)
}
\arguments{
\item{RES}{List with simulation results}

\item{n_quantiles}{}
}
\value{

}
\description{
Plot the patience distribution
}
