% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pltQueueByHourPerc}
\alias{pltQueueByHourPerc}
\title{Analyze the customers by patience during an arrival cycle}
\usage{
pltQueueByHourPerc(RES, n_quantiles = 4)
}
\arguments{
\item{RES}{List with simulation results}

\item{n_quantiles}{}
}
\value{

}
\description{
Analyze the customers by patience during an arrival cycle
}
