% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pltQueueLengthArrivals}
\alias{pltQueueLengthArrivals}
\title{Plot First Arrivals and Queue Length Changes}
\usage{
pltQueueLengthArrivals(RES, n_customers = 500, pch = 4)
}
\arguments{
\item{RES}{List with simulation results}

\item{n_customers}{no. of first customers to plot}

\item{pch}{what character to use for arrivals}
}
\value{

}
\description{
Plot First Arrivals and Queue Length Changes
}
