% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RATE}
\alias{RATE}
\title{Rate function factory}
\usage{
RATE(gamma, lambda_0)
}
\arguments{
\item{lambda_0}{}
}
\value{
A function with argument t that computes the arrival rate
}
\description{
Rate function factory
}
\examples{
rate <- RATE(5,10)
rate(1:3)
}
