% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{readAWXFiles}
\alias{readAWXFiles}
\title{Title Read all simulation files in a folder}
\usage{
readAWXFiles(path = getwd())
}
\arguments{
\item{path}{The folder path. Defaults to the current working directory.}
}
\value{
a list with elemets comprising the AWX tagbles.
}
\description{
Title Read all simulation files in a folder
}
\examples{
L <- readAWXFiles()
}
