% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{resSimCosine}
\alias{resSimCosine}
\title{Atomic Simulation for periodic arrivals (cosine model)}
\usage{
resSimCosine(n, gamma, lambda_0, theta, s, eta, mu)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{gamma}{The periodic arrival rate maximum amplitude}

\item{lambda_0}{The constant arrival rate}

\item{theta}{Parameter of the exponential patience}

\item{s}{Number of servers.}

\item{eta}{Shape parameter of job size.}

\item{mu}{Rate parameter of job size.}
}
\value{
A list with the estimates
}
\description{
Atomic Simulation for periodic arrivals (cosine model)
}
