% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{VW}
\alias{VW}
\title{Liron's Virtual Waiting function}
\usage{
VW(Res.service, s)
}
\arguments{
\item{Res.service}{the vector of residual service times}

\item{s}{the number of servers}
}
\description{
Liron's Virtual Waiting function
}
