% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{negLogLikelihoodMean.KnownArrival}
\alias{negLogLikelihoodMean.KnownArrival}
\title{(negative) mean log-likelihood for known sinusoidal arrivals
The mean is returned instead of the sum - should help gradient based optimizers}
\usage{
negLogLikelihoodMean.KnownArrival(theta, params, dati)
}
\arguments{
\item{theta}{a vector of theta values}

\item{params}{A vector with values (gamma,lambda_0)}

\item{dati}{compact simulation results for memory economy (A,W,X).}
}
\value{
The negative log-likelihood at the point provided.
}
\description{
(negative) mean log-likelihood for known sinusoidal arrivals
The mean is returned instead of the sum - should help gradient based optimizers
}
